% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef_components.R
\name{bef_components}
\alias{bef_components}
\title{Makes a local population projection and produce results for population
components}
\usage{
bef_components(startpop, assumptions, YEAR)
}
\arguments{
\item{startpop}{This is the start population}

\item{assumptions}{This is a Data Frame with assumptions}

\item{YEAR}{This is the year from which the forecast starts}
}
\value{
The output from \code{\link{return}}
}
\description{
Makes a local population projection and produce results for population
components
}
\examples{
bef_components(startpop_data,assump_data,2019)
}
