% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef_proj.R
\name{bef_proj}
\alias{bef_proj}
\title{Makes a local population projection and produce results for growth per year.}
\usage{
bef_proj(startpop, assumptions, YEAR)
}
\arguments{
\item{startpop}{This is the start population}

\item{assumptions}{This is a Data Frame with assumptions}

\item{YEAR}{This is the year from which the forecast starts}
}
\value{
The output from \code{\link{return}}
}
\description{
Makes a local population projection and produce results for growth per year.
}
\examples{
bef_proj(startpop_data,assump_data,2019)
}
