% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef_raw.R
\name{bef_raw}
\alias{bef_raw}
\title{Makes a local population projection and produce results for age, sex and
year}
\usage{
bef_raw(startpop, assumptions, YEAR)
}
\arguments{
\item{startpop}{This is the start population}

\item{assumptions}{This is a Data Frame with assumptions}

\item{YEAR}{This is the year from which the forecast starts}
}
\value{
The output from \code{\link{return}}
}
\description{
Makes a local population projection and produce results for age, sex and
year
}
\examples{
bef_raw(startpop_data,assump_data,2019)
}
